componentsディレクトリに移動
