+++ src/index.tsx
@@ -1,23 +1,8 @@
 import { Options as FocusTrapOptions } from 'focus-trap';
-import React, { useCallback, useMemo, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
+import React, { useCallback, useMemo, useState } from 'react';
 
-import { useBodyScrollLock } from './hooks/useBodyScrollLock';
-import { useFocusTrap } from './hooks/useFocusTrap';
-
-export interface ModalProps {
-  children: React.ReactNode;
-  isOpen: boolean;
-  close: () => void;
-  elementId: 'root' | string;
-  preventScroll: boolean;
-  focusTrapOptions: FocusTrapOptions;
-  closeButton: React.ReactElement | null;
-}
-
-interface DefaultCloseButtonProps {
-  onClose: () => void;
-}
+import { DefaultCloseButton } from './components/DefaultCloseButton';
+import { Modal, ModalProps } from './components/Modal';
 
 export type ModalOptions = {
   preventScroll?: boolean;
@@ -41,105 +26,6 @@ export type UseModal = (
   close: () => void,
   isOpen: boolean
 ];
-
-const wrapperStyle: React.CSSProperties = {
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  bottom: 0,
-  right: 0,
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  zIndex: 1000,
-};
-
-const overlayStyle: React.CSSProperties = {
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  bottom: 0,
-  right: 0,
-  backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  zIndex: 100000,
-};
-
-const containerStyle: React.CSSProperties = {
-  position: 'relative',
-  zIndex: 100001,
-};
-
-const closeButtonStyle: React.CSSProperties = {
-  // reset
-  backgroundColor: 'transparent',
-  border: 'none',
-  cursor: 'pointer',
-  outline: 'none',
-  padding: 0,
-  appearance: 'none',
-
-  position: 'absolute',
-  right: 0,
-  top: 0,
-  width: '40px',
-  height: '40px',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  fontSize: '20px',
-};
-
-const DefaultCloseButton: React.FC<DefaultCloseButtonProps> = ({ onClose }) => {
-  return (
-    <button
-      type="button"
-      style={closeButtonStyle}
-      onClick={onClose}
-      aria-label="close"
-    >
-      ×
-    </button>
-  );
-};
-
-const Modal: React.FC<ModalProps> = ({
-  children,
-  isOpen,
-  close,
-  elementId = 'root',
-  preventScroll,
-  focusTrapOptions,
-  closeButton,
-}) => {
-  const dialogRef = useRef<HTMLDivElement>(null);
-  useFocusTrap(dialogRef, isOpen, {
-    onDeactivate: close,
-    clickOutsideDeactivates: true,
-    ...focusTrapOptions,
-  });
-  useBodyScrollLock(dialogRef, isOpen, preventScroll);
-
-  if (isOpen === false) {
-    return null;
-  }
-
-  return createPortal(
-    <div style={wrapperStyle}>
-      <div style={overlayStyle} />
-      <div
-        ref={dialogRef}
-        role="dialog"
-        aria-modal="true"
-        style={containerStyle}
-        tabIndex={-1}
-      >
-        {children}
-        {closeButton}
-      </div>
-    </div>,
-    document.getElementById(elementId) as HTMLElement
-  );
-};
 
 export const useModal: UseModal = (elementId = 'root', options = {}) => {
   const { preventScroll = false, focusTrapOptions = {}, ...rest } = options;
